app: add image slider
